fix: Avoid truncating last char from folded strings
As described in #433, if the last character of a folded string is the
very last character in the file (no trailing whitespace), it's not
included the token's `Value`.

While fixing this, I also noticed that the column information for the
folded and literal strings were the end rather than the beginning. I've
(hopefully) fixed that as well.

Fixes #433

Signed-off-by: Charith Ellawala <charith@cerbos.dev>